fix: composer
