+++ composer.json
@@ -86,6 +86,7 @@
     },
     "require-dev": {
         "beyondcode/laravel-dump-server": "^2.1",
+        "brianium/paratest": "^7.10",
         "fakerphp/faker": "^1.9.1",
         "kitloong/laravel-migrations-generator": "^7.0",
         "larastan/larastan": "^3.0",
